added comment section as a bonus feature

+++ sessions/recap-project-6/example-implementation/pages/places/[id]/index.js
@@ -2,6 +2,7 @@ import Link from "next/link";
 import { useRouter } from "next/router.js";
 import useSWR from "swr";
 import styled from "styled-components";
+import Comments from "../../../components/Comments.js";
 import { StyledLink } from "../../../components/StyledLink.js";
 import { StyledButton } from "../../../components/StyledButton.js";
 import { StyledImage } from "../../../components/StyledImage.js";
@@ -38,16 +39,19 @@ export default function DetailsPage() {
   if (!isReady || isLoading || error) return <h2>Loading...</h2>;
 
   async function deletePlace() {
-    await fetch(`/api/places/${id}`, {
-      method: "DELETE",
-    });
-    router.push("/");
+    if (confirm("are you sure you want to delete this place?")) {
+      await fetch(`/api/places/${id}`, {
+        method: "DELETE",
+      });
+      router.push("/");
+      return;
+    }
   }
 
   return (
     <>
       <Link href={"/"} passHref legacyBehavior>
-        <StyledLink $justifySelf="start">back</StyledLink>
+        <StyledLink justifySelf="start">back</StyledLink>
       </Link>
       <ImageContainer>
         <StyledImage
@@ -71,10 +75,11 @@ export default function DetailsPage() {
         <Link href={`/places/${id}/edit`} passHref legacyBehavior>
           <StyledLink>Edit</StyledLink>
         </Link>
-        <StyledButton onClick={deletePlace} type="button" $variant="delete">
+        <StyledButton onClick={deletePlace} type="button" variant="delete">
           Delete
         </StyledButton>
       </ButtonContainer>
+      <Comments locationName={place.name} />
     </>
   );
 }
